class03-20250427/05_litellm/litellm-basic/pyproject.toml
[project]
name = "litellm-basic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Hassan Ali Khan", email = "73381969+hassan-ak@users.noreply.github.com" }
]
requires-python = ">=3.11"
dependencies = [
    "litellm>=1.67.2",
    "python-dotenv>=1.1.0",
]

[project.scripts]
gemini = "litellm_basic.main:run_gemini"
groq = "litellm_basic.main:run_groq"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
